app-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-aiplatform
python-dotenv